Fixed warnings made by compiler & clang-tidy.

+++ src/models/ticket.hpp
@@ -302,31 +302,31 @@ public:
         size_t n=0;
         sql +="default,";
         if(dirtyFlag_[1]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[2]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[3]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[4]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[5]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[6]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[7]) {
-            n = sprintf(placeholderStr, "$%d,", placeholder++);
+            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
             sql.append(placeholderStr, n);
         }
         if(parametersCount > 0)
